backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
httpx
python-dotenv